+++ index.html
@@ -8,7 +8,7 @@
     <title>Document</title>
     <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
     <link rel="stylesheet" href="style.css">
-    <script src="main.js" defer type="module"></script>
+    <script src="index.js" defer type="module"></script>
 </head>
 
 <body>
