+++ modules/Valuables.js
@@ -1,0 +1,7 @@
+export const titleInput = document.querySelector('.title');
+export const authorInput = document.querySelector('.author');
+export const form = document.querySelector('form');
+export const booksListElement = document.querySelector('.listOfBooks');
+export const navigationLinks = document.querySelectorAll('.nav-link');
+export const timeIndicator = document.querySelector('.timeIndicator');
+
